More refactoring. Rewrote image resize/convert and operator tests. Server tests still needs to be rewritten. Server and all features work now.

+++ server/ops/liquid_rescale.go
@@ -3,11 +3,11 @@ package ops
 import "github.com/phzfi/RIC/server/images"
 
 type LiquidRescale struct {
-	Width, Height uint
+	Width, Height int
 }
 
 func (r LiquidRescale) Apply(img images.Image) error {
 	// The third argument to LiquidRescaleImage is the maximum transversal step, or how many pixels a seam can move sideways.
 	// The fourth is the rigidity, which makes seams less steep. Recommended if transversal step is greater than one.
-	return img.LiquidRescaleImage(r.Width, r.Height, 1, 0)
+	return img.LiquidRescaleImage(uint(r.Width), uint(r.Height), 1, 0)
 }
